% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Function to aggregate data for youth vs. adult comparison}
\usage{
aggregate_data(data)
}
\arguments{
\item{data}{A .csv dataframe containing data to be aggregated.}
}
\value{
nothing; but changes the given dataframe.
}
\description{
Function to aggregate data for youth vs. adult comparison
}
