% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{create_bar_use_plot}
\alias{create_bar_use_plot}
\title{Drug Use Data EDA Functions:}
\usage{
create_bar_use_plot(
  data,
  x_var,
  y_var,
  title,
  x_label,
  y_label,
  fill_color,
  output_file
)
}
\arguments{
\item{data}{A .csv dataframe containing data to be plotted.}

\item{x_var}{A string specifying the x-axis variable.}

\item{y_var}{A string specifying the y-axis variable.}

\item{title}{A string specifying the plot title.}

\item{x_label}{A string specifying the x-axis label.}

\item{y_label}{A string specifying the y-axis label.}

\item{fill_color}{A string specifying the fill color for the bars.}

\item{output_file}{A string specifying the output file path for saving the plot.}
}
\value{
ggplot objects representing the generated plot
}
\description{
Functions to run exploratory data analysis on a drug use dataset:
\itemize{
\item \code{create_bar_use_plot()}: Creates a bar plot for substance use.
\item \code{create_bar_freq_plot()}: Creates a bar plot for substance use frequency.
\item \code{create_scatter_plot()}: Creates a scatter plot with a regression line.
\item \code{aggregate_data()}: Aggregates data for youth vs. adult comparison.
\item \code{create_grouped_bar_plot()}: Creates grouped bar plots for youth vs. adult comparison.
Function to create bar plot for substance use
}
}
\examples{
sample_data <- data.frame(
 age = c("18-24", "25-34", "35-44"),
 `alcohol-use` = c(80, 75, 60),
 `alcohol-frequency` = c(20, 15, 10),
 `marijuana-use` = c(50, 30, 20),
 `heroin-frequency` = c(5, 3, 2),
 class = c("youth", "adult", "adult"),
 n = c(100, 200, 150)
)
create_bar_use_plot(sample_data, "age", "alcohol-use", "Alcohol Use", "Age", "Proportion", "dodgerblue", "output/eda-test/test1.png")
}
