% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory.R
\name{create_directory}
\alias{create_directory}
\title{Create directory if it doesn't exist}
\usage{
create_directory(path)
}
\arguments{
\item{path}{A character string specifying the directory path.}
}
\value{
None. The function creates a directory as a side effect.
}
\description{
Checks if the specified directory exists, and
creates the directory path if necessary.
}
