% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{create_grouped_bar_plot}
\alias{create_grouped_bar_plot}
\title{Function to create grouped bar plots}
\usage{
create_grouped_bar_plot(data, title, x_label, y_label, output_file)
}
\arguments{
\item{data}{A .csv dataframe containing data to be plotted or aggregated.}

\item{title}{A string specifying the plot title.}

\item{x_label}{A string specifying the x-axis label.}

\item{y_label}{A string specifying the y-axis label.}

\item{output_file}{A string specifying the output file path for saving the plot.}
}
\value{
ggplot objects representing the generated plot.
}
\description{
Function to create grouped bar plots
}
