% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Function to create scatter plot with regression line}
\usage{
create_scatter_plot(
  data,
  x_var,
  y_var,
  color_var,
  title,
  x_label,
  y_label,
  output_file
)
}
\arguments{
\item{data}{A .csv dataframe containing data to be plotted.}

\item{x_var}{A string specifying the x-axis variable.}

\item{y_var}{A string specifying the y-axis variable.}

\item{color_var}{A string specifying the color variable for the scatter plot.}

\item{title}{A string specifying the plot title.}

\item{x_label}{A string specifying the x-axis label.}

\item{y_label}{A string specifying the y-axis label.}

\item{output_file}{A string specifying the output file path for saving the plot.}
}
\value{
ggplot objects representing the generated plot
}
\description{
Function to create scatter plot with regression line
}
\examples{
sample_data <- data.frame(
 age = c("18-24", "25-34", "35-44"),
 `alcohol-use` = c(80, 75, 60),
 `alcohol-frequency` = c(20, 15, 10),
 `marijuana-use` = c(50, 30, 20),
 `heroin-frequency` = c(5, 3, 2),
 class = c("youth", "adult", "adult"),
 n = c(100, 200, 150)
)
create_scatter_plot(sample_data, "alcohol-frequency", "heroin-frequency", "age", "Relationship Between Alcohol and Heroin Frequency Use", "Alcohol Median Frequency", "Heroin Median Frequency", "output/eda-test/test5.png")
}
