% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download Dataset}
\usage{
download_data(url, destination)
}
\arguments{
\item{url}{A character string specifying the file URL.}

\item{destination}{A character string specifying the
file path where the data will be saved.}
}
\value{
None. The function downloads the file as a side effect.
}
\description{
Downloads a file from a given URL and saves it to a specified destination.
}
\examples{
if (interactive()) {
  test_url <- paste0("https://raw.githubusercontent.com/rudeboybert/",
  "fivethirtyeight/master/data-raw/drug-use-by-age/drug-use-by-age.csv")
  test_dir <- "data/test"
  download_data(test_url, test_dir)
}
}
