% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{save_data}
\alias{save_data}
\title{Save processed data to a file}
\usage{
save_data(data, output_path)
}
\arguments{
\item{data}{A data frame to be saved.}

\item{output_path}{A string specifying the output file path.}
}
\value{
Nothing. Saves the data to a file.
}
\description{
Save processed data to a file
}
\examples{
\dontrun{
save_data(mtcars, "data/clean/mtcars-cleaned.csv")
}
}
