% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_and_evaluate.R
\name{train_and_evaluate}
\alias{train_and_evaluate}
\title{Train and evaluate a model, generate confusion matrix, and save results}
\usage{
train_and_evaluate(
  model_spec,
  model_name,
  data_train,
  data_test,
  recipe,
  output_path
)
}
\arguments{
\item{model_spec}{A tidymodels model specification.}

\item{model_name}{A string representing the model name (used for saving outputs).}

\item{data_train}{A data frame containing the training data.}

\item{data_test}{A data frame containing the testing data.}

\item{recipe}{A tidymodels recipe for preprocessing the data.}

\item{output_path}{A string specifying the directory to save results.}
}
\description{
This function fits a model, generates a confusion matrix, and saves the results as both a PNG and CSV.
}
