% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Function to aggregate data for youth vs. adult comparison}
\usage{
aggregate_data(data)
}
\arguments{
\item{data}{A .csv dataframe containing data to be aggregated.}
}
\value{
nothing; but changes the given dataframe.
}
\description{
Function to aggregate data for youth vs. adult comparison
}
\examples{
sample_data <- data.frame(
 age = c("18-24", "25-34", "35-44"),
 alcohol.use = c(80, 75, 60),
 alcohol.frequency = c(20, 15, 10),
 marijuana.use = c(50, 30, 20),
 heroin.frequency = c(5, 3, 2),
 class = c("youth", "adult", "adult"),
 n = c(100, 200, 150)
)
aggregate_data(sample_data)
}
