% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{create_bar_freq_plot}
\alias{create_bar_freq_plot}
\title{Create and save frequency bar plots}
\usage{
create_bar_freq_plot(
  data,
  x_var,
  y_var,
  title,
  x_label,
  y_label,
  fill_color,
  output_file
)
}
\arguments{
\item{data}{A .csv dataframe containing data to be plotted.}

\item{x_var}{A string specifying the x-axis variable.}

\item{y_var}{A string specifying the y-axis variable.}

\item{title}{A string specifying the plot title.}

\item{x_label}{A string specifying the x-axis label.}

\item{y_label}{A string specifying the y-axis label.}

\item{fill_color}{A string specifying the fill color for the bars.}

\item{output_file}{A string specifying the output file path for saving the plot.}
}
\value{
ggplot objects representing the generated plot
}
\description{
Create and save frequency bar plots
}
