% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-functions.R
\name{create_bar_use_plot}
\alias{create_bar_use_plot}
\title{Using the cleaned data, create some exploratory data analysis plots to gain a better understanding of
the data.}
\usage{
create_bar_use_plot(
  data,
  x_var,
  y_var,
  title,
  x_label,
  y_label,
  fill_color,
  output_file
)
}
\arguments{
\item{data}{A .csv dataframe containing data to be plotted or aggregated.}

\item{x_var}{A string specifying the x-axis variable.}

\item{y_var}{A string specifying the y-axis variable.}

\item{title}{A string specifying the plot title.}

\item{x_label}{A string specifying the x-axis label.}

\item{y_label}{A string specifying the y-axis label.}

\item{fill_color}{A string specifying the fill color for the bars.}

\item{output_file}{A string specifying the output file path for saving the plot.}

\item{color_var}{A string specifying the color variable for the scatter plot.}
}
\value{
ggplot objects representing the generated plot(s).
}
\description{
These functions create different types of plots based on the scope of certain variables
in the data. The functions include:
\itemize{
\item \code{create_bar_use_plot}: Creates a bar plot for substance use.
\item \code{create_bar_freq_plot}: Creates a bar plot for substance use frequency.
\item \code{create_scatter_plot}: Creates a scatter plot with a regression line.
\item \code{aggregate_data}: Aggregates data for youth vs. adult comparison.
\item \code{create_grouped_bar_plot}: Creates grouped bar plots for youth vs. adult comparison.
}
}
