% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{create_grouped_bar_plot}
\alias{create_grouped_bar_plot}
\title{Function to create grouped bar plots}
\usage{
create_grouped_bar_plot(data, title, x_label, y_label, output_file)
}
\arguments{
\item{data}{A .csv dataframe containing data to be plotted or aggregated.}

\item{title}{A string specifying the plot title.}

\item{x_label}{A string specifying the x-axis label.}

\item{y_label}{A string specifying the y-axis label.}

\item{output_file}{A string specifying the output file path for saving the plot.}
}
\value{
ggplot objects representing the generated plot.
}
\description{
Function to create grouped bar plots
}
\examples{
sample_aggregated <- tibble(
 class = c("adult", "adult", "adult", "adult", "adult", "youth", "youth", "youth", "youth", "youth"),
 total_n = c(350, 350, 350, 350, 350, 100, 100, 100, 100, 100),
 variable = c("alcohol-use", "alcohol-frequency", "marijuana-use", "heroin-frequency", "n", "alcohol-use", "alcohol-frequency", "marijuana-use", "heroin-frequency", "n"),
 value = c(68.6, 12.9, 25.7, 2.57, 179., 80, 20, 50, 5, 100),
)
create_grouped_bar_plot(sample_aggregated, "Youth vs. Adult Comparison", "Substance Type", "Mean Substance Use (\%)", "output/eda-test/test6.png")
}
