% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Function to create scatter plot with regression line}
\usage{
create_scatter_plot(
  data,
  x_var,
  y_var,
  color_var,
  title,
  x_label,
  y_label,
  output_file
)
}
\arguments{
\item{data}{A .csv dataframe containing data to be plotted.}

\item{x_var}{A string specifying the x-axis variable.}

\item{y_var}{A string specifying the y-axis variable.}

\item{color_var}{A string specifying the color variable for the scatter plot.}

\item{title}{A string specifying the plot title.}

\item{x_label}{A string specifying the x-axis label.}

\item{y_label}{A string specifying the y-axis label.}

\item{output_file}{A string specifying the output file path for saving the plot.}
}
\value{
ggplot objects representing the generated plot
}
\description{
Function to create scatter plot with regression line
}
